# openapi/koru_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "koru",
    "description": "An API for managing long-term shared expenses.",
    "contact": {
      "name": "Raphael Biland",
      "email": "raphael.biland@outlook.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.4.0"
  },
  "paths": {
...
    "/groups/{group_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Fetches a group, with it's un-settled expenses.",
        "description": "Fetches a group, with it's un-settled expenses.\n\nThis action can only be performed by a group member.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExamples (replace GROUP_ID ith the group UUID):\n```\ncurl -i -b cookie \"http://localhost:8000/groups/GROUP_ID\"\n```\n",
        "operationId": "get_group",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Deletes a group.",
        "description": "Deletes a group.\n\nThis action can only be performed by the group administrator.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExample (replace GROUP_ID ith the group UUID):\n```\ncurl -i -b cookie -X DELETE \"http://localhost:8000/groups/GROUP_ID\"\n```\n",
        "operationId": "delete_group",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Group deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/groups/{group_id}/settlements": {
      "get": {
        "tags": [
          "Settlements"
        ],
        "summary": "Fetches all settlements of the group.",
        "description": "Fetches all settlements of the group.\n\nThis action can only be performed by a group member.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExamples (replace GROUP_ID ith the group UUID):\n```\ncurl -i -b cookie \"http://localhost:8000/groups/GROUP_ID/settlements\"\n```\n",
        "operationId": "get_settlements",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of settlements of the group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettlementsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Settlements"
        ],
        "summary": "Settles the current expenses and returns the settlement summary.",
        "description": "Settles the current expenses and returns the settlement summary.\n\nThis action can only be performed by the group administrator.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExample (replace GROUP_ID ith the group UUID):\n```\ncurl -i -H 'Content-Type: application/json' -b cookie -X POST \"http://localhost:8000/groups/GROUP_ID/settlements\"\n```\n",
        "operationId": "settle",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Settlement computed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs in the user given valid credentials.",
        "description": "Logs in the user given valid credentials.\n\nIt will return a `Set-Cookie` header, that should be used in next authenticated requests\n\nExample:\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"password\":\"123\",\"email\":\"r@r1.com\"}' -c cookie \"http://localhost:8000/login\"\n```",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Auth cookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors in login request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColorDto": {
        "type": "object",
        "required": [
          "red",
          "green",
          "blue"
        ],
        "properties": {
          "blue": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "green": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "red": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          }
        }
      },
...
      "DetailedGroupDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "members",
          "expenses"
        ],
        "properties": {
          "expenses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseDto"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDto"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorData": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ErrorData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ExpenseDto": {
        "type": "object",
        "required": [
          "id",
          "description",
          "amount",
          "user",
          "date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/MemberDto"
          }
        }
      },
...
      "GetGroupResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupResponse"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GetSettlementsResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SettlementsResponse"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GroupResponse": {
        "type": "object",
        "required": [
          "group"
        ],
        "properties": {
          "group": {
            "$ref": "#/components/schemas/DetailedGroupDto"
          }
        }
      },
...
      "IdData": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "IdResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IdData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "LoginPayload": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MemberDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "is_admin",
          "color",
          "joined_at"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_admin": {
            "type": "boolean"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "MessageData": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "SettlementDto": {
        "type": "object",
        "required": [
          "id",
          "end_date",
          "transactions"
        ],
        "properties": {
          "end_date": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDto"
            }
          }
        }
      },
      "SettlementResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SettlementDto"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "SettlementsResponse": {
        "type": "object",
        "required": [
          "settlements"
        ],
        "properties": {
          "settlements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettlementDto"
            }
          }
        }
      },
      "TransactionDto": {
        "type": "object",
        "required": [
          "from",
          "to",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "float"
          },
          "from": {
            "$ref": "#/components/schemas/MemberDto"
          },
          "to": {
            "$ref": "#/components/schemas/MemberDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id"
      }
    }
  },
...
}